+++ examples/python/fill_and_fair.py
@@ -1,6 +1,7 @@
 import os
 import sys
-from tinymesh import Mesh, hole_fill, remesh_incremental
+import numpy as np
+from tinymesh import Mesh, hole_fill, remesh_incremental, implicit_fair
 
 
 def main(filename):
@@ -12,15 +13,22 @@ def main(filename):
         f.set_is_static(True)
 
     # Fill holes
-    hole_fill(mesh)
+    hole_fill(mesh, np.pi / 6.0)
+    print('[ OK ] hole filling')
 
     # Then, remesh
-    remesh_incremental(mesh)
+    remesh_incremental(mesh, short_length=0.5, long_length=2.0)
+    print('[ OK ] remeshing')
+
+    # Finally, fair
+    implicit_fair(mesh)
+    print('[ OK ] fairing')
 
     # Save
     base, ext = os.path.splitext(filename)
     outfile = base + "_remesh" + ext
     mesh.save(outfile)
+    print('[ OK ] saved to %s' % (outfile))
 
 
 if __name__ == '__main__':