+++ CHANGELOG.md
@@ -2,3 +2,6 @@ Master
 ---
 
 *   Added a user option to `setup.py` to take CMake build arguments from command line.
+*   Implicit fairing `implicit_fair` is updated to its normalized version.
+*   Hole filling `hole_fill` is updated to take the upper bound of allowed dihedral angle.
+*   Update `setup.py` to support the operation `install`.